// src/defaults/DefaultActions.d.ts
declare const DefaultActions: {
    DEFAULT: number;
    PRIMARY: number;
    SECONDARY: number;
    FORWARD: number;
    BACKWARD: number;
    UP: number;
    DOWN: number;
    LEFT: number;
    RIGHT: number;
    INTERACT: number;
    CROUCH: number;
    JUMP: number;
    WALK: number;
    RUN: number;
    SPRINT: number;
    SNEAK: number;
};
export default DefaultActions;